PROJECT_garbage_classifier/requirements.txt
torch
torchvision
numpy
matplotlib
seaborn
scikit-learn
tqdm
Pillow
tabulate
sys
time
json
os